Some small fixes, cleanup and refactoring Relates #25 Relates #26
